Update tiffany project structure
We create a pipeline package rather than putting everything inside cmd package
