+++ cmd/root.go
@@ -1,3 +1,7 @@
+// Copyright 2019 Thinking Machines Data Science. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root
+// for license information.
+
 // Package cmd contains all the helper functions, handlers, and command-line methods
 // for building the tiffany command-line interface.
 package cmd
@@ -9,6 +13,8 @@ import (
 
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
+
+	"github.com/thinkingmachines/tiffany/pkg/pipeline"
 )
 
 var rootCmd = &cobra.Command{
@@ -47,49 +53,13 @@ var rootCmd = &cobra.Command{
 		// Get arguments passed
 		coordinate := []string{args[0], args[1]}
 		initLogger(verbosity)
-		client := GetStaticMapsClient(env)
-		skip := RunPipeline(client, coordinate, zoom, size, path, noRef, wtLbl, force)
+		client := pipeline.GetStaticMapsClient(env)
+		skip := pipeline.Run(client, coordinate, zoom, size, path, noRef, wtLbl, force)
 		log.WithFields(log.Fields{
 			"lat":     coordinate[0],
 			"lon":     coordinate[1],
 			"skipped": skip,
 		}).Info("Single job done!")
-	},
-}
-
-var batchCmd = &cobra.Command{
-	Use:   "batch PATH/TO/FILE.CSV",
-	Short: "Apply tiffany on a CSV file of coordinates",
-	Long: `
-The batch command is a more efficient alternative when running tiffany
-on a list of lat-lon coordinates. Instead of using a for-loop, you can
-just provide the path to the CSV file, and apply the same parameters as
-if you're running tiffany on a single point.
-
-Assumes that the first column is the latitude and the second column is the
-longitude.
-`,
-	Example: `
-  tiffany batch coordinates.csv
-  tiffany batch coordinates.csv --without-reference
-  tiffany batch coordinates.csv --with-labels=/path/to/file.shp
-`,
-	Args:    cobra.ExactArgs(1),
-	Version: "v1.0.0-alpha.2",
-	Run: func(cmd *cobra.Command, args []string) {
-		// Get arguments passed
-		csvFile := args[0]
-		log.WithFields(log.Fields{
-			"file": csvFile,
-		}).Info("Batch job successfully started")
-		initLogger(verbosity)
-		client := GetStaticMapsClient(env)
-		total, numSkip := RunBatchPipeline(client, csvFile, skipFirst, zoom, size, path, noRef, wtLbl, force)
-		fmt.Println("")
-		log.WithFields(log.Fields{
-			"total":   total,
-			"skipped": numSkip,
-		}).Info("Batch job done!")
 	},
 }
 
@@ -121,14 +91,6 @@ func init() {
 	batchCmd.Flags().BoolVar(&skipFirst, "skip-header-row", false, "skip header row")
 }
 
-// Execute runs the root command
-func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
-
 func initLogger(verbosity int) {
 	if verbosity == 1 {
 		log.SetLevel(log.DebugLevel)
@@ -136,5 +98,13 @@ func initLogger(verbosity int) {
 		log.SetLevel(log.TraceLevel)
 	} else {
 		log.SetLevel(log.InfoLevel)
+	}
+}
+
+// Execute runs the root command
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
 	}
 }
