+++ main.go
@@ -1,3 +1,7 @@
+// Copyright 2019 Thinking Machines Data Science. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root
+// for license information.
+
 /*
 Tiffany is command-line tool for rendering to TIFF any image from Google Static Maps.
